Tratamento de response

+++ api_rest/views.py
@@ -17,21 +17,34 @@ def get_users(request):
     if request.method == 'GET':
         users = Usuario.objects.all()
         serializer = UsuarioSerializer(users, many=True)
-        return Response(serializer.data)
-    return Response(status=status.HTTP_400_BAD_REQUEST)
+        formatted_response = {
+            'success': True,
+            'message': 'Lista de usuários recuperada com sucesso.',
+            'data': serializer.data
+        }
+        return Response(formatted_response, status=status.HTTP_200_OK)
+    return Response({'error': 'Requisição inválida'}, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['GET'])
 def get_by_id(request, id):
-
     try: 
         user = Usuario.objects.get(pk=id)
     except:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+        formatted_response = {
+            'success': False,
+            'message': 'Usuário não localizado.'
+        }
+        return Response(formatted_response, status=status.HTTP_404_NOT_FOUND)
     
     if request.method == 'GET':
         serializer = UsuarioSerializer(user)
-        return Response(serializer.data)
+        formatted_response = {
+            'success': True,
+            'message': 'Usuário recuperado com sucesso.',
+            'data': serializer.data
+        }
+        return Response(formatted_response, status=status.HTTP_200_OK)
     
 
 @api_view(['POST'])
@@ -42,7 +55,12 @@ def post_create_user(request):
 
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            formatted_response = {
+            'success': True,
+            'message': 'Usuário criado com sucesso.',
+            'data': serializer.data
+        }
+            return Response(formatted_response, status=status.HTTP_201_CREATED)
         return Response(status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -51,22 +69,34 @@ def put_edit_user(request, id):
     try:
         update_user = Usuario.objects.get(pk=id)
     except Usuario.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+        formatted_response = {
+            'success': False,
+            'message': 'Usuário não localizado.'
+        }
+        return Response(formatted_response, status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'PUT':
         serializer = UsuarioSerializer(update_user, data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_202_ACCEPTED)
+            formatted_response = {
+            'success': True,
+            'message': 'Usuário editado com sucesso.',
+            'data': serializer.data
+        }
+            return Response(formatted_response, status=status.HTTP_202_ACCEPTED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-
 @api_view(['DELETE'])
-def delete_deleted_user(request, id):  
+def delete_user(request, id):
     try:
-        if request.method == 'DELETE':  
-            Usuario.objects.filter(id=id).delete()
-            return Response(status=status.HTTP_200_OK)
-    except Usuario.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)   
-        
+        if request.method == 'DELETE':
+            user_exists = Usuario.objects.filter(id=id).exists()
+            if user_exists:
+                Usuario.objects.filter(id=id).delete()
+                return JsonResponse({'success': True, 'message': 'Usuário deletado com sucesso.'}, status=status.HTTP_200_OK)
+            else:
+                return JsonResponse({'success': False, 'message': 'Usuário não localizado.'}, status=status.HTTP_404_NOT_FOUND)
+    
+    except:
+        return JsonResponse({'success': False, 'message': 'Erro interno no servidor:'}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
